Binarizarea imaginii cu algoritmul triunghiului

+++ Algorithms/Tools/Tools.cs
@@ -66,11 +66,30 @@ namespace Algorithms.Tools
 
         #region Thresholding
 
-        public static Image<Gray, byte> Thresholding(Image<Gray, byte> inputImage, byte treshold)
+        public static Image<Gray, byte> Thresholding(Image<Gray, byte> inputImage, int treshold)
         {
             Image<Gray, byte> emguCvImg = inputImage.ThresholdBinary(new Gray(treshold), new Gray(255));
 
             return emguCvImg;
+
+            //Image<Gray, byte> result = new Image<Gray, byte>(inputImage.Size);
+
+            //for (int y = 0; y < inputImage.Height; y++)
+            //{
+            //    for (int x = 0; x < inputImage.Width; x++)
+            //    {
+            //        if (inputImage.Data[y, x, 0] >= treshold)
+            //        {
+            //            result.Data[y, x, 0] = 255;
+            //        }
+            //        else
+            //        {
+            //            result.Data[y, x, 0] = 0;
+            //        }
+            //    }
+            //}
+
+            //return result;
         }
 
         #endregion
